moved training loop into the module
