+++ pyproject.toml
@@ -3,7 +3,7 @@ target-version = "py311"
 line-length = 119
 fix = true
 lint.select = ["E","F","W","B"]
-lint.ignore = ["E501", "F841"]
+lint.ignore = ["E501", "F841", "F811"]
 
 [tool.pytest.ini_options]
 testpaths = ["smtb/tests"]
